SkinPrediction/backend/requirements.txt
Flask==3.0.0
tensorflow==2.15.0
numpy==1.26.2
Pillow==10.1.0
flask-cors==4.0.0
gunicorn==21.2.0
flask-sqlalchemy==3.1.1
python-dotenv==1.0.0
Werkzeug==3.0.1
scikit-learn==1.3.2
matplotlib==3.8.2
seaborn==0.13.0
